refactor: minor test utils refactoring
